Lift slots state up to App
